random number functions hotfix

+++ DatabaseEngine.cs
@@ -84,7 +84,7 @@ namespace UnlockDB
             engine.SetValue("md5", new Func<string, string>(UnlockDB.Helpers.ScriptUtils.MD5));
             engine.SetValue("sha256", new Func<string, string>(UnlockDB.Helpers.ScriptUtils.SHA256));
             engine.SetValue("toString", new Func<object, string>(UnlockDB.Helpers.ScriptUtils.ToString));
-            engine.SetValue("randomNumber", new Func<double, double, double>(UnlockDB.Helpers.ScriptUtils.RandomNumber));
+            engine.SetValue("randomNumber", new Func<int, int, int>(UnlockDB.Helpers.ScriptUtils.RandomNumber));
             engine.SetValue("randomDecimal", new Func<string, string, decimal>(UnlockDB.Helpers.ScriptUtils.RandomDecimal));
             engine.SetValue("randomString", new Func<int, string>(UnlockDB.Helpers.ScriptUtils.RandomString));
             engine.SetValue("toBase64", new Func<string, string>(UnlockDB.Helpers.ScriptUtils.ToBase64));
